memory_system/requirements.txt
# Core Dependencies
dataclasses-json>=0.6.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Database
sqlite3  # Built-in with Python

# Optional Database Backends
# postgresql: psycopg2-binary>=2.9.0
# mongodb: pymongo>=4.0.0
# redis: redis>=4.0.0

# Framework Dependencies (install as needed)

# LangChain
# langchain>=0.1.0
# langchain-community>=0.0.10
# langchain-openai>=0.0.5

# LangGraph
# langgraph>=0.0.20

# CrewAI
# crewai>=0.1.0

# AutoGen
# pyautogen>=0.2.0

# LlamaIndex
# llama-index>=0.9.0

# Agno
# agno>=0.1.0  # If available

# Embeddings (optional)
# openai>=1.0.0
# sentence-transformers>=2.2.0
# faiss-cpu>=1.7.0

# Vector Databases (optional)
# pinecone-client>=2.2.0
# weaviate-client>=3.20.0
# qdrant-client>=1.5.0
# chromadb>=0.4.0

# Utilities
# numpy>=1.24.0
# pandas>=2.0.0
# tiktoken>=0.5.0  # For token counting

# Development Dependencies
# pytest>=7.4.0
# pytest-cov>=4.1.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# mypy>=1.5.0
# flake8>=6.1.0

# Documentation
# sphinx>=7.2.0
# sphinx-rtd-theme>=1.3.0
# autodoc>=0.5.0
